+++ tests/blob/apis/blob.test.ts
@@ -1001,11 +1001,11 @@ describe("BlobAPIs", () => {
     const result_copy = await destBlobClient.syncCopyFromURL(
       sourceBlobClient.url
     );
-
     assert.equal(
-      result_copy!._response.request.headers.get("x-ms-client-request-id"),
-      result_copy!._response.request.requestId
+      result_copy._response.request.headers.get("x-ms-client-request-id"),
+      result_copy._response.request.requestId
     );
+    assert.equal(result_copy.copyStatus, "success");
 
     const result = await destBlobClient.getProperties();
     assert.ok(result.date);
@@ -1119,9 +1119,10 @@ describe("BlobAPIs", () => {
       sourceBlobClient.url
     );
     assert.equal(
-      result_copy!._response.request.headers.get("x-ms-client-request-id"),
-      result_copy!._response.request.requestId
+      result_copy._response.request.headers.get("x-ms-client-request-id"),
+      result_copy._response.request.requestId
     );
+    assert.equal(result_copy.copyStatus, "success");
 
     const result = await destBlobClient.getProperties();
     assert.ok(result.date);